requirements.txt
streamlit
openai
python-dotenv
pdfplumber
